fix: build error
```
cmd/main.go:145:25: undefined: InitDefaultKeystore
cmd/main.go:166:2: undefined: CheckLockError
cmd/main.go:283:3: undefined: CheckLockError
```

need to do more refactoring
